Removed training script and improved documentation
