+++ ts/torch_handler/text_classifier.py
@@ -5,14 +5,17 @@ Module for text classification default handler
 DOES NOT SUPPORT BATCH!
 """
 import logging
+
 import torch
 import torch.nn.functional as F
-from torchtext.data.utils import ngrams_iterator
 from captum.attr import TokenReferenceBase
-from .text_handler import TextHandler
+from torchtext.data.utils import ngrams_iterator
+
 from ..utils.util import map_class_to_label
+from .text_handler import TextHandler
 
 logger = logging.getLogger(__name__)
+
 
 class TextClassifier(TextHandler):
     """
@@ -45,8 +48,9 @@ class TextClassifier(TextHandler):
 
         line = data[0]
         text = line.get("data") or line.get("body")
+        # Decode text if not a str but bytes or bytearray
         if isinstance(text, (bytes, bytearray)):
-            text = text.decode('utf-8')
+            text = text.decode("utf-8")
 
         text = self._remove_html_tags(text)
         text = text.lower()
@@ -55,11 +59,8 @@ class TextClassifier(TextHandler):
         text = self._remove_punctuation(text)
         text = self._tokenize(text)
         text_tensor = torch.as_tensor(
-            [
-                self.source_vocab[token]
-                for token in ngrams_iterator(text, self.ngrams)
-            ],
-            device=self.device
+            [self.source_vocab[token] for token in ngrams_iterator(text, self.ngrams)],
+            device=self.device,
         )
         return text_tensor, text
